refactor: move loader related files into its own directory
